Restructured test directory. Mocked out files for the graph functionality.
